requirements.txt
wheel
aiohttp
aiofiles
aria2p
async_generator
cowpy
psutil
wget
telethon-tgcrypto
cffi>=1.0.0
dnspython
emoji
PyPDF2
cryptg
patool
gitpython
asyncurban
google-api-python-client==1.8.0
google-auth-oauthlib
googletrans==3.1.0a0
google_images_download>=2.7.1
gTTS>=2.1.1
gTTS-token>=1.1.3
hachoir>=3.0a5
REPO_LINK
heroku3>=3.4.0	
httplib2>=0.18.0
humanize 
lxml
oauth2client
Pillow>=6.0.0
fontTools
psycopg2
psycopg2-binary
pybase64>=0.4.0
pylast
pySmartDL
python-barcode
python-dotenv
pytz
qrcode
requests>=2.18.4
search-engine-parser>=0.6.1
speedtest-cli>=2.0.2
sqlalchemy<=1.3.23
telethon>=1.14.0 
telethon-session-sqlalchemy>=0.2.6
urbandict>=0.5
wikipedia>=1.4.0
coffeehouse>=2.1.0
git+https://github.com/johnwmillr/LyricsGenius.git
git+https://github.com/voins/google-images-download.git@2cd68173c961324a2c41c61e8b6f40a49663ce60
lyricsgenius
redis>=3.3.10
PyDrive2
pyDownload
beautifulsoup4>=4.8.1
PyGithub
bs4
covid>=2.2.9
base
pip>=19.2.3
setuptools>=28.8.0
selenium>=3.141.0 
telegraph
python-dateutil
pyfiglet   
jotquote
html_telegraph_poster
csvfaker
git+https://github.com/LonamiWebs/Telethon
git+https://github.com/ahupp/python-magic
pytube>=9.5.1
pymongo[srv]
youtube_search
justwatch
deezloader
zipfile38
bitlyshortener
validators
geopy
glitch_this
twitter_scraper
pyaztro
jikanpy
vcsi
lottie
wand
colour
scipy
wordcloud
pendulum
opencv-python
hentai
natsort
yt-dlp
youtube-dl>=2021.1.8
youtube-search-python>=1.3.2
youtube-search>=1.1.1
